Attempting to add a file on the new VM
I was able to create the VM. But not move the file over. I think I need to add a connection block.

A strange thing that happened was the main.tf built the resources. But when I went to destroy it told me the password value was not correct? WTF I went into the portal and removed them.