Bastion and no to Provisioners
Today I learned:
Using Provisioners is not recommended to add files after a VM is built. It seems in a bind they can work, but there are better ways. Instead, I should use custom_data on Azure and user_data on aws. On to learning how that works.

So I'm still stuck on getting the file over after the VM is provisioned.

I added a Bastion resource to the build for access to the VM after creation. I enjoy the idea of being able to access the VM's without a public IP. Though it seems the provider azurerm_bastion_host needs a Public IP in the build. If this is true, then maybe I don't understand Bastion as well as I would like, which is probably true.

One step at a time, things do happen.